fix non-inclusion of generated META-INF/services files
When the META-INF/services files did not previously exist, any assembly
of the apk would lose the SPI declarations because they were generated
too late in the build toolchain to be 'merged'. The solution was to set
each variant's mergeResources task to be dependent upon the tasks which
generate the SPI declarations.